+++ src/chi-vue/public/index.html
@@ -5,12 +5,12 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width,initial-scale=1.0">
     <title>Chi Vue</title>
-    <link rel="stylesheet" href="http://localhost:8000/chi.css">
-    <link rel="icon" type="image/svg+xml" href="http://localhost:8000/assets/images/favicon.svg">
-    <link rel="alternate icon" href="http://localhost:8000/assets/images/favicon.ico">
-    <script src="http://localhost:8000/js/chi.js"></script>
-    <script type="module" src="https://assets.ctl.io/chi/4.1.1/js/ce/ux-chi-ce/ux-chi-ce.esm.js"></script>
-    <script nomodule="" src="https://assets.ctl.io/chi/4.1.1/js/ce/ux-chi-ce/ux-chi-ce.js"></script>
+    <link rel="stylesheet" href="https://assets.ctl.io/chi/4.3.0/chi.css">
+    <link rel="icon" type="image/svg+xml" href="https://assets.ctl.io/chi/4.3.0/assets/images/favicon.svg">
+    <link rel="alternate icon" href="https://assets.ctl.io/chi/4.3.0/assets/images/favicon.ico">
+    <script src="https://assets.ctl.io/chi/4.3.0/js/chi.js"></script>
+    <script type="module" src="https://assets.ctl.io/chi/4.3.0/js/ce/ux-chi-ce/ux-chi-ce.esm.js"></script>
+    <script nomodule="" src="https://assets.ctl.io/chi/4.3.0/js/ce/ux-chi-ce/ux-chi-ce.js"></script>
   </head>
   <body>
     <noscript>
