+++ src/chi-vue/src/components/data-table/DataTable.tsx
@@ -29,6 +29,7 @@ import Tooltip from '../tooltip/tooltip';
 import Pagination from '../pagination/pagination';
 import DataTableToolbar from '@/components/data-table-toolbar/DataTableToolbar';
 import arraySort from 'array-sort';
+import { defaultConfig } from './default-config';
 
 let dataTableNumber = 0;
 @Component({})
@@ -37,10 +38,12 @@ export default class DataTable extends Vue {
   @Prop() config!: DataTableConfig;
 
   accordionsExpanded: string[] = [];
-  currentPage = this.config.activePage || 1;
-  resultsPerPage = this.config.resultsPerPage || 20;
+  activePage =
+    this.$props.config.pagination.activePage || this.$props.config.activePage || defaultConfig.pagination.activePage;
+  resultsPerPage = this.$props.config.resultsPerPage || defaultConfig.resultsPerPage;
   selectedRows: string[] = [];
   slicedData: DataTableRow[] = [];
+  mode = this.$props.config.mode || defaultConfig.mode;
   _currentScreenBreakpoint?: DataTableScreenBreakpoints;
   _dataTableId?: string;
   _expandable!: boolean;
@@ -312,6 +315,13 @@ export default class DataTable extends Vue {
   }
 
   _calculateNumberOfPages() {
+    const serverSide = this.$props.config.mode === 'serverside';
+    const pages = this.$props.config.pagination.pages;
+
+    if (serverSide && pages && typeof pages === 'number') {
+      return pages;
+    }
+
     return Math.max(Math.ceil(this.data.body.length / this.resultsPerPage), 1);
   }
 
@@ -580,7 +590,10 @@ export default class DataTable extends Vue {
   _pagination() {
     const pages = this._calculateNumberOfPages();
 
-    if ((pages === 1 && this.config.pagination.hideOnSinglePage) || this.$props.data.body.length === 0) {
+    if (
+      (pages === 1 && this.config.pagination.hideOnSinglePage) ||
+      (this.$props.data.body.length === 0 && this.$props.config.mode !== 'serverside')
+    ) {
       return null;
     } else {
       return (
@@ -589,7 +602,7 @@ export default class DataTable extends Vue {
             ref="pagination"
             compact={this.config.style.portal || this.config.pagination.compact}
             firstLast={this.config.pagination.firstLast}
-            currentPage={this.currentPage}
+            currentPage={this.activePage}
             pages={pages}
             results={this.data.body.length}
             pageSize={!this.config.style.portal}
@@ -618,8 +631,8 @@ export default class DataTable extends Vue {
 
   sliceData(data: DataTableRow[]): DataTableRow[] {
     if (data.length > this.resultsPerPage) {
-      const arrayStartIndex = (this.currentPage - 1) * this.resultsPerPage,
-        arrayEndIndex = (this.currentPage - 1) * this.resultsPerPage + this.resultsPerPage;
+      const arrayStartIndex = (this.activePage - 1) * this.resultsPerPage,
+        arrayEndIndex = (this.activePage - 1) * this.resultsPerPage + this.resultsPerPage;
 
       return data.slice(arrayStartIndex, arrayEndIndex);
     }
@@ -682,13 +695,18 @@ export default class DataTable extends Vue {
       if (rowObject.expanded && !this.accordionsExpanded.includes(rowObject.rowId)) {
         this.accordionsExpanded.push(rowObject.rowId);
       }
+      if (rowObject.selected && !this.selectedRows.includes(rowObject.rowId)) {
+        this.selectedRows.push(rowObject.rowId);
+      }
       return rowObject;
     };
     let rowNumber = 0;
 
     this._serializedDataBody = [];
-    // eslint-disable-next-line
-    this._expandable = !!this.data.body.find((row: { nestedContent: any }) => row.nestedContent);
+    this._expandable =
+      this.$props.config.reserveExpansionSlot ||
+      // eslint-disable-next-line
+      !!this.data.body.find((row: { nestedContent: any }) => row.nestedContent);
     this.data.body.forEach(row => {
       this._serializedDataBody.push(serializeRow(row, rowNumber, null));
       rowNumber++;
@@ -723,6 +741,11 @@ export default class DataTable extends Vue {
     this.slicedData = this.sliceData(this._sortedData || this._serializedDataBody);
   }
 
+  @Watch('config')
+  dataconfigChange() {
+    this.activePage = this.$props.config.pagination.activePage || 1;
+  }
+
   created() {
     this._dataTableId = `dt-${dataTableNumber}`;
     dataTableNumber += 1;
@@ -749,6 +772,7 @@ export default class DataTable extends Vue {
 
   mounted() {
     const dataTableComponent = this.$refs.dataTable as HTMLElement;
+
     if (dataTableComponent && this.config.columnResize) {
       this._generateColumnResize(dataTableComponent);
     }
@@ -765,14 +789,17 @@ export default class DataTable extends Vue {
       (this.$refs.pagination as Vue).$on(PAGINATION_EVENTS.PAGE_CHANGE, (ev: number) => {
         const data = this._sortedData && this._sortedData.length > 0 ? this._sortedData : this._serializedDataBody;
         const numberOfPages = this._calculateNumberOfPages();
-        const pageChangeEventData: DataTablePageChange = {
-          page: ev,
-          data: this.slicedData,
-        };
 
         if (ev >= 1 && ev <= numberOfPages) {
-          this.currentPage = ev;
-          this.slicedData = this.sliceData(data);
+          const pageChangeEventData: DataTablePageChange = {
+            page: ev,
+          };
+
+          if (this.mode === 'clientside') {
+            this.activePage = ev;
+            this.slicedData = this.sliceData(data);
+            pageChangeEventData.data = this.slicedData;
+          }
           this.$emit(PAGINATION_EVENTS.PAGE_CHANGE, pageChangeEventData);
           this._checkSelectAllCheckbox();
         }
@@ -888,14 +915,17 @@ export default class DataTable extends Vue {
           const sortingData: DataTableSorting = {
             column: columnName,
             direction: 'descending',
-            data: this._sortedData,
           };
+
+          if (this.mode === 'clientside') {
+            this.sortData(columnName, 'descending', columnSortBy);
+            sortingData.data = this._sortedData;
+          }
 
           (sortIcon as HTMLElement).style.transform = 'rotate(180deg)';
           columnHeadSortButton.setAttribute('data-sort', 'descending');
-          this.sortData(columnName, 'descending', columnSortBy);
           this.$emit(DATA_TABLE_EVENTS.DATA_SORTING, sortingData);
-          this.currentPage = 1;
+          this.activePage = 1;
           if (this._sortedData) {
             this.slicedData = this.sliceData(this._sortedData);
           }
@@ -906,27 +936,42 @@ export default class DataTable extends Vue {
             sortBy: columnSortBy || columnName,
           };
         } else if (currentSort === 'descending' && this._sortConfig && this._sortConfig.key === columnName) {
+          const sortingData: DataTableSorting = {
+            column: undefined,
+            direction: undefined,
+          };
+
+          if (this.mode === 'clientside') {
+            sortingData.data = undefined;
+          }
+
           sortIcon.className = `${ICON_CLASS} -xs ${DATA_TABLE_SORT_ICONS.SORT}`;
           columnHeadSortButton.removeAttribute('data-sort');
           (sortIcon as HTMLElement).removeAttribute('style');
           columnHeadSortButton.blur();
-          if (this._sortedData) {
-            this._sortedData.length = 0;
+          if (this.$props.config.mode !== 'serverside') {
+            if (this._sortedData) {
+              this._sortedData.length = 0;
+            }
+            this.slicedData = this.sliceData(this._serializedDataBody);
           }
-          this.slicedData = this.sliceData(this._serializedDataBody);
+          this.$emit(DATA_TABLE_EVENTS.DATA_SORTING, sortingData);
           columnHeadCell?.classList.remove(ACTIVE_CLASS);
           this._sortConfig = undefined;
         } else {
-          const sortingData = {
+          const sortingData: DataTableSorting = {
             column: columnName,
             direction: 'ascending',
-            sortedData: this._sortedData,
           };
 
+          if (this.mode === 'clientside') {
+            this.sortData(columnName, 'ascending', columnSortBy);
+            sortingData.data = this._sortedData;
+          }
+
           columnHeadSortButton.setAttribute('data-sort', 'ascending');
-          this.sortData(columnName, 'ascending', columnSortBy);
           this.$emit(DATA_TABLE_EVENTS.DATA_SORTING, sortingData);
-          this.currentPage = 1;
+          this.activePage = 1;
           if (this._sortedData) {
             this.slicedData = this.sliceData(this._sortedData);
           }
