+++ src/chi-vue/src/components/data-table/default-config.ts
@@ -1,0 +1,25 @@
+import { DataTableConfig } from '@/constants/types';
+
+export const defaultConfig: DataTableConfig = {
+  activePage: 1,
+  columnResize: false,
+  mode: 'clientside',
+  noResultsMessage: 'No matches found. Please revise search criteria and try again.',
+  pagination: {
+    compact: false,
+    firstLast: false,
+    hideOnSinglePage: false,
+    pageJumper: false,
+  },
+  resultsPerPage: 10,
+  style: {
+    portal: false,
+    compact: false,
+    bordered: false,
+    noBorder: false,
+    hover: true,
+    striped: false,
+    size: 'md',
+  },
+  selectable: false,
+};
