--- src/chi-vue/src/views/DataTableView.vue
@@ -1,1295 +1,0 @@
-<template>
-  <div id="tadataTableVIew">
-    <h2>Data Table example</h2>
-    <ChiDataTable
-      :data="table"
-      :config="config"
-      @chiSelectedRowsChange="e => this.rowSelect(e)"
-      @chiPageChange="e => this.pageChange(e)"
-      @chiPageSizeChange="e => this.pageSizeChange(e)"
-      @chiDataSorting="e => this.dataSorting(e)"
-      @chiRowExpanded="e => this.rowExpanded(e)"
-      @chiRowCollapsed="e => this.rowCollapsed(e)"
-    >
-      <template #icon="payload">
-        <i :class="`chi-icon icon-${payload.icon} -icon--${payload.color}`"></i>
-      </template>
-      <template #ticketId="payload">
-        <TicketPopover :id="payload.id" />
-      </template>
-      <template #status="payload">
-        <div :class="`chi-badge ${payload.status === 'active' ? '-primary' : ''}`">
-          <span class="-text--truncate">{{ payload.status }}</span>
-        </div>
-      </template>
-      <template #actions="payload">
-        <Actions :id="payload.id" />
-      </template>
-      <template #accordionContent="payload">
-        <div class="chi-alert -success" role="alert">
-          <i class="chi-alert__icon chi-icon icon-circle-check"></i>
-          <div class="chi-alert__content">
-            <p class="chi-alert__text">
-              Custom content rendered by the provided template. ID:
-              <b>{{ payload.id }}</b>
-            </p>
-          </div>
-        </div>
-      </template>
-      <template #date="payload">
-        {{ `${payload.date.getDate()} ${months[payload.date.getMonth()]} ${payload.date.getFullYear()}` }}
-      </template>
-      <template #toolbar>
-        <ChiDataTableToolbar
-          @chiToolbarFiltersChange="e => filtersChange(e)"
-          @chiToolbarSearch="e => search(e)"
-          @chiToolbarColumnsChange="e => columnsChange(e)"
-        >
-          <ChiSearchInput size="lg" :dataTableSearch="true" />
-          <ChiDataTableFilters :filtersData="toolbar.filtersData" class="-ml--2" />
-          <ChiColumnCustomization :columnsData="toolbar.columnsData" />
-        </ChiDataTableToolbar>
-      </template>
-      <template #loadingSkeleton>
-        <div class="-d--flex -flex--column -w--100">
-          <div class="chi-skeleton -w--85 -w-md--75 -w-lg--50"></div>
-          <div class="chi-skeleton -xs -w--90 -w-lg--70 -mt--2"></div>
-          <div class="chi-skeleton -xs -w--95 -w-lg--80 -mt--1"></div>
-          <div class="chi-skeleton -xs -w--55 -w-lg--55 -mt--1"></div>
-        </div>
-      </template>
-    </ChiDataTable>
-  </div>
-</template>
-
-<script lang="ts">
-import { Component, Vue } from 'vue-property-decorator';
-import DataTable from '../components/data-table/DataTable';
-import Actions from './DataTableTemplates/example-actions.vue';
-import TicketPopover from './DataTableTemplates/example-popover.vue';
-import DataTableToolbar from '../components/data-table-toolbar/DataTableToolbar';
-import SearchInput from '../components/search-input/SearchInput';
-import DataTableFilters from '../components/data-table-filters/DataTableFilters';
-import { DataTableRow } from '../constants/types';
-import ColumnCustomization from '../components/column-customization/ColumnCustomization';
-
-@Component({
-  components: {
-    ChiDataTable: DataTable,
-    ChiDataTableToolbar: DataTableToolbar,
-    ChiSearchInput: SearchInput,
-    ChiDataTableFilters: DataTableFilters,
-    ChiColumnCustomization: ColumnCustomization,
-    Actions,
-    TicketPopover,
-  },
-  methods: {
-    rowSelect: e => {
-      console.log(e);
-    },
-    pageChange: e => {
-      console.log(e);
-    },
-    pageSizeChange: e => {
-      console.log(e);
-    },
-    dataSorting: e => {
-      console.log(e);
-    },
-    search: e => {
-      console.log(e);
-    },
-    filtersChange: e => {
-      console.log(e);
-    },
-    rowExpanded: e => {
-      console.log(`Expanded row: `, e);
-    },
-    rowCollapsed: e => {
-      console.log(`Collapsed row: `, e);
-    },
-  },
-  data: () => {
-    return {
-      config: {
-        columnResize: true,
-        noResultsMessage: 'No matches found. Please revise search criteria and try again.',
-        activePage: 1,
-        style: {
-          portal: true,
-          noBorder: false,
-          bordered: false,
-          hover: false,
-          size: 'lg',
-          striped: true,
-        },
-        pagination: {
-          hideOnSinglePage: true,
-          compact: true,
-          firstLast: true,
-          pageJumper: true,
-        },
-        selectable: true,
-        columnSizes: {
-          sm: [10, 10, 10, 0, 10, 10, 10, 5],
-          md: [15, 10, 15, 0, 15, 15, 15, 5],
-          lg: [15, 10, 15, 0, 15, 15, 15, 5],
-          xl: [15, 10, 10, 15, 15, 15, 15, 5],
-        },
-        resultsPerPage: 10,
-        defaultSort: {
-          key: 'ticketId',
-          sortBy: 'id',
-          direction: 'ascending',
-        },
-      },
-      toolbar: {
-        filtersData: [
-          {
-            name: 'status',
-            type: 'select',
-            options: [
-              {
-                label: 'Active',
-                value: 'active',
-              },
-              {
-                label: 'Inactive',
-                value: 'inactive',
-              },
-            ],
-            value: 'active',
-          },
-          {
-            name: 'input',
-            placeholder: 'Input filter',
-            type: 'input',
-            value: '',
-          },
-          {
-            name: 'checkbox',
-            type: 'checkbox',
-            checked: false,
-          },
-          {
-            name: 'statusAdvanced',
-            label: 'Status',
-            type: 'select',
-            options: [
-              {
-                label: 'Active',
-                value: 'active',
-                selected: false,
-              },
-              {
-                label: 'Inactive',
-                value: 'inactive',
-                selected: false,
-              },
-            ],
-            advanced: true,
-            value: 'active',
-          },
-          {
-            name: 'inputAdvanced',
-            label: 'Label',
-            type: 'input',
-            advanced: true,
-          },
-          {
-            name: 'textareaAdvanced',
-            label: 'textarea',
-            type: 'textarea',
-            advanced: true,
-          },
-          {
-            name: 'checkboxAdvanced',
-            label: 'Checkbox',
-            type: 'checkbox',
-            advanced: true,
-          },
-        ],
-        columnsData: {
-          columns: [
-            {
-              name: 'columnA',
-              label: 'Column A',
-            },
-            {
-              name: 'columnB',
-              label: 'Column B',
-            },
-            {
-              name: 'columnC',
-              label: 'Column C',
-              locked: true,
-              selected: true,
-            },
-            {
-              name: 'columnD',
-              label: 'Column D',
-              locked: true,
-              selected: true,
-            },
-            {
-              name: 'columnE',
-              label: 'Column E',
-              selected: true,
-            },
-            {
-              name: 'columnF',
-              label: 'Column F',
-              selected: true,
-            },
-            {
-              name: 'columnG',
-              label: 'Column G',
-            },
-            {
-              name: 'columnH',
-              label: 'Column H',
-            },
-            {
-              name: 'columnI',
-              label: 'Column I',
-            },
-            {
-              name: 'columnJ',
-              label: 'Column J',
-            },
-          ],
-        },
-      },
-      table: {
-        head: {
-          ticketId: { label: 'Ticket ID', sortable: true, sortBy: 'id', sortDataType: 'string', key: true, bold: true },
-          alerts: { label: 'Alerts', align: 'center' },
-          status: {
-            label: 'Status',
-            sortable: true,
-            sortBy: 'status',
-            sortDataType: 'string',
-            align: 'center',
-          },
-          string: { label: 'String', sortable: true, sortDataType: 'string', key: true },
-          number: { label: 'Number', sortable: true, sortDataType: 'number' },
-          date: { label: 'Date', sortable: true, sortBy: 'date', sortDataType: 'date' },
-          productId: { label: 'Product / Service ID', sortable: true, sortDataType: 'string' },
-          actions: { label: ' ', align: 'right', allowOverflow: true },
-        },
-        body: [
-          {
-            id: 'NTM000021063',
-            nestedContent: {
-              template: 'loadingSkeleton',
-            },
-            active: false,
-            data: [
-              { template: 'ticketId', payload: { id: 'NTM000021063' } },
-              {
-                template: 'icon',
-                payload: { icon: 'circle-check', color: 'success' },
-              },
-              { template: 'status', payload: { status: 'active' } },
-              'Colocation A',
-              0,
-              { template: 'date', payload: { date: new Date('02/05/2018 8:00 AM') } },
-              'SVUJW034781A',
-              {
-                template: 'actions',
-                payload: { id: 'NTM000021063' },
-                align: 'right',
-              },
-            ],
-          },
-          {
-            id: 'NTM000021071',
-            active: false,
-            data: [
-              { template: 'ticketId', payload: { id: 'NTM000021071' } },
-              {},
-              { template: 'status', payload: { status: 'inactive' } },
-              'Colocation B',
-              1,
-              { template: 'date', payload: { date: new Date('03/05/2018 8:00 AM') } },
-              'SVUJW034781A',
-              {
-                template: 'actions',
-                payload: { id: 'NTM000021071' },
-                align: 'right',
-              },
-            ],
-          },
-          {
-            id: 'NTM000021064',
-            active: false,
-            nestedContent: {
-              value: 'asdasdasd',
-            },
-            data: [
-              { template: 'ticketId', payload: { id: 'NTM000021064' } },
-              '',
-              { template: 'status', payload: { status: 'inactive' } },
-              'Colocation B',
-              2,
-              { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-              'SVUJW034781A',
-              {
-                template: 'actions',
-                payload: { id: 'NTM000021064' },
-                align: 'right',
-              },
-            ],
-          },
-          {
-            id: 'NTM000021065',
-            active: false,
-            data: [
-              { template: 'ticketId', payload: { id: 'NTM000021065' } },
-              {
-                label: 'Alert Column',
-                template: 'icon',
-                payload: { icon: 'warning', color: 'warning' },
-              },
-              { template: 'status', payload: { status: 'active' } },
-              'Colocation C',
-              3,
-              { template: 'date', payload: { date: new Date('05/05/2018 8:00 AM') } },
-              'SVUJW034781A',
-              {
-                template: 'actions',
-                payload: { id: 'NTM000021065' },
-                align: 'right',
-              },
-            ],
-          },
-          {
-            id: 'NTM000021066',
-            active: false,
-            nestedContent: {
-              template: 'accordionContent',
-              payload: {
-                id: 'NTM000021066',
-              },
-            },
-            data: [
-              { template: 'ticketId', payload: { id: 'NTM000021066' } },
-              '',
-              { template: 'status', payload: { status: 'active' } },
-              'Colocation a',
-              4,
-              { template: 'date', payload: { date: new Date('06/05/2018 8:00 AM') } },
-              'SVUJW034781A',
-              {
-                template: 'actions',
-                payload: { id: 'NTM000021066' },
-                align: 'right',
-              },
-            ],
-          },
-          {
-            id: 'NTM0000210662',
-            active: false,
-            nestedContent: {
-              table: {
-                data: [
-                  {
-                    id: 'NTM0000210661',
-                    active: false,
-                    data: [
-                      {
-                        template: 'ticketId',
-                        payload: { id: 'NTM0000210661' },
-                      },
-                      {
-                        label: 'Alerts',
-                        template: 'icon',
-                        payload: { icon: 'circle-check', color: 'success' },
-                      },
-                      {
-                        template: 'status',
-                        payload: { status: 'active' },
-                        align: 'center',
-                      },
-                      'Colocation',
-                      5,
-                      { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-                      'SVUJW034781A',
-                      {
-                        template: 'actions',
-                        payload: { id: 'NTM0000210661' },
-                        align: 'right',
-                      },
-                    ],
-                  },
-                  {
-                    id: 'NTM00002106611',
-                    active: false,
-                    nestedContent: {
-                      table: {
-                        data: [
-                          {
-                            id: 'a',
-                            active: false,
-                            data: [
-                              {
-                                template: 'ticketId',
-                                payload: { id: 'NTM0000210661' },
-                              },
-                              {
-                                label: 'Alerts',
-                                template: 'icon',
-                                payload: {
-                                  icon: 'circle-check',
-                                  color: 'success',
-                                },
-                              },
-                              {
-                                template: 'status',
-                                payload: { status: 'active' },
-                                align: 'center',
-                              },
-                              'Colocation',
-                              6,
-                              { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-                              'SVUJW034781A',
-                              {
-                                template: 'actions',
-                                payload: { id: 'NTM0000210661' },
-                                align: 'right',
-                              },
-                            ],
-                          },
-                          {
-                            id: 'bbb',
-                            active: false,
-                            nestedContent: {
-                              table: {
-                                data: [
-                                  {
-                                    id: 'ba',
-                                    active: false,
-                                    data: [
-                                      {
-                                        template: 'ticketId',
-                                        payload: { id: 'NTM00002106611' },
-                                      },
-                                      {
-                                        label: 'Alerts',
-                                        template: 'icon',
-                                        payload: {
-                                          icon: 'circle-check',
-                                          color: 'success',
-                                        },
-                                      },
-                                      {
-                                        template: 'status',
-                                        payload: { status: 'active' },
-                                        align: 'center',
-                                      },
-                                      'Colocation',
-                                      7,
-                                      { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-                                      'SVUJW034781A',
-                                      {
-                                        template: 'actions',
-                                        payload: { id: 'NTM00002106611' },
-                                        align: 'right',
-                                      },
-                                    ],
-                                  },
-                                  {
-                                    id: 'bb',
-                                    active: false,
-                                    data: [
-                                      {
-                                        template: 'ticketId',
-                                        payload: { id: 'NTM00002106621' },
-                                      },
-                                      {
-                                        label: 'Alerts',
-                                        template: 'icon',
-                                        payload: {
-                                          icon: 'circle-check',
-                                          color: 'success',
-                                        },
-                                      },
-                                      {
-                                        template: 'status',
-                                        payload: { status: 'active' },
-                                        align: 'center',
-                                      },
-                                      'Colocation',
-                                      8,
-                                      { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-                                      'SVUJW034781A',
-                                      {
-                                        template: 'actions',
-                                        payload: { id: 'NTM00002106621' },
-                                        align: 'right',
-                                      },
-                                    ],
-                                  },
-                                ],
-                              },
-                            },
-                            data: [
-                              {
-                                template: 'ticketId',
-                                payload: { id: 'NTM0000210662' },
-                              },
-                              {
-                                label: 'Alerts',
-                                template: 'icon',
-                                payload: {
-                                  icon: 'circle-check',
-                                  color: 'success',
-                                },
-                              },
-                              {
-                                template: 'status',
-                                payload: { status: 'active' },
-                                align: 'center',
-                              },
-                              'Colocation',
-                              9,
-                              { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-                              'SVUJW034781A',
-                              {
-                                template: 'actions',
-                                payload: { id: 'NTM0000210662' },
-                                align: 'right',
-                              },
-                            ],
-                          },
-                        ],
-                      },
-                    },
-                    data: [
-                      {
-                        template: 'ticketId',
-                        payload: { id: 'NTM0000210662' },
-                      },
-                      {
-                        label: 'Alerts',
-                        template: 'icon',
-                        payload: { icon: 'circle-check', color: 'success' },
-                      },
-                      {
-                        template: 'status',
-                        payload: { status: 'active' },
-                        align: 'center',
-                      },
-                      'Colocation',
-                      10,
-                      { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-                      'SVUJW034781A',
-                      {
-                        template: 'actions',
-                        payload: { id: 'NTM0000210662' },
-                        align: 'right',
-                      },
-                    ],
-                  },
-                ],
-              },
-            },
-            data: [
-              { template: 'ticketId', payload: { id: 'NTM000021067' } },
-              '',
-              { template: 'status', payload: { status: 'inactive' } },
-              'Colocation b',
-              11,
-              { template: 'date', payload: { date: new Date('07/05/2018 8:00 AM') } },
-              'SVUJW034781A',
-              {
-                template: 'actions',
-                payload: { id: 'NTM000021067' },
-                align: 'right',
-              },
-            ],
-          },
-          {
-            id: 'NTM000021072',
-            active: false,
-            data: [
-              { template: 'ticketId', payload: { id: 'NTM000021072' } },
-              {},
-              { template: 'status', payload: { status: 'active' } },
-              'Colocation c',
-              12,
-              { template: 'date', payload: { date: new Date('01/06/2018 8:00 AM') } },
-              'SVUJW034781A',
-              {
-                template: 'actions',
-                payload: { id: 'NTM000021072' },
-                align: 'right',
-              },
-            ],
-          },
-          {
-            id: 'NTM0000210651',
-            active: false,
-            data: [
-              { template: 'ticketId', payload: { id: 'NTM0000210651' } },
-              {
-                label: 'Alerts',
-                template: 'icon',
-                payload: { icon: 'warning', color: 'warning' },
-              },
-              { template: 'status', payload: { status: 'active' } },
-              'Colocation',
-              13,
-              { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-              'SVUJW034781A',
-              {
-                template: 'actions',
-                payload: { id: 'NTM0000210651' },
-                align: 'right',
-              },
-            ],
-          },
-          {
-            id: 'NTM000021068',
-            active: false,
-            data: [
-              { template: 'ticketId', payload: { id: 'NTM000021068' } },
-              {},
-              { template: 'status', payload: { status: 'active' } },
-              'Colocation',
-              14,
-              { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-              'SVUJW034781A',
-              {
-                template: 'actions',
-                payload: { id: 'NTM000021068' },
-                align: 'right',
-              },
-            ],
-          },
-          {
-            id: 'NTM000021069',
-            active: false,
-            data: [
-              { template: 'ticketId', payload: { id: 'NTM000021069' } },
-              {},
-              { template: 'status', payload: { status: 'active' } },
-              'Colocation',
-              15,
-              { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-              'SVUJW034781A',
-              {
-                template: 'actions',
-                payload: { id: 'NTM000021069' },
-                align: 'right',
-              },
-            ],
-          },
-          {
-            id: 'NTM000021070',
-            active: false,
-            data: [
-              { template: 'ticketId', payload: { id: 'NTM000021070' } },
-              {},
-              { template: 'status', payload: { status: 'inactive' } },
-              'Colocation',
-              16,
-              { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-              'SVUJW034781A',
-              {
-                template: 'actions',
-                payload: { id: 'NTM000021070' },
-                align: 'right',
-              },
-            ],
-          },
-          {
-            id: 'NTM000021073',
-            active: false,
-            data: [
-              { template: 'ticketId', payload: { id: 'NTM000021073' } },
-              {},
-              { template: 'status', payload: { status: 'inactive' } },
-              'Colocation',
-              17,
-              { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-              'SVUJW034781A',
-              {
-                template: 'actions',
-                payload: { id: 'NTM000021073' },
-                align: 'right',
-              },
-            ],
-          },
-          {
-            id: 'NTM000021074',
-            active: false,
-            data: [
-              { template: 'ticketId', payload: { id: 'NTM000021074' } },
-              {},
-              { template: 'status', payload: { status: 'inactive' } },
-              'Colocation',
-              18,
-              { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-              'SVUJW034781A',
-              {
-                template: 'actions',
-                payload: { id: 'NTM000021074' },
-                align: 'right',
-              },
-            ],
-          },
-          {
-            id: 'NTM000021075',
-            active: false,
-            data: [
-              { template: 'ticketId', payload: { id: 'NTM000021075' } },
-              {},
-              { template: 'status', payload: { status: 'active' } },
-              'Colocation',
-              19,
-              { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-              'SVUJW034781A',
-              {
-                template: 'actions',
-                payload: { id: 'NTM000021075' },
-                align: 'right',
-              },
-            ],
-          },
-          {
-            id: 'NTM000021076',
-            active: false,
-            data: [
-              { template: 'ticketId', payload: { id: 'NTM000021076' } },
-              {},
-              { template: 'status', payload: { status: 'inactive' } },
-              'Colocation',
-              20,
-              { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-              'SVUJW034781A',
-              {
-                template: 'actions',
-                payload: { id: 'NTM000021076' },
-                align: 'right',
-              },
-            ],
-          },
-          {
-            id: 'NTM000021077',
-            active: false,
-            data: [
-              { template: 'ticketId', payload: { id: 'NTM000021077' } },
-              {},
-              { template: 'status', payload: { status: 'active' } },
-              'Colocation',
-              21,
-              { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-              'SVUJW034781A',
-              {
-                template: 'actions',
-                payload: { id: 'NTM000021077' },
-                align: 'right',
-              },
-            ],
-          },
-          {
-            id: 'NTM000021078',
-            active: false,
-            data: [
-              { template: 'ticketId', payload: { id: 'NTM000021078' } },
-              {},
-              { template: 'status', payload: { status: 'active' } },
-              'Colocation',
-              22,
-              { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-              'SVUJW034781A',
-              {
-                template: 'actions',
-                payload: { id: 'NTM000021078' },
-                align: 'right',
-              },
-            ],
-          },
-          {
-            id: 'NTM000021079',
-            active: false,
-            data: [
-              { template: 'ticketId', payload: { id: 'NTM000021079' } },
-              {},
-              { template: 'status', payload: { status: 'active' } },
-              'Colocation',
-              23,
-              { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-              'SVUJW034781A',
-              {
-                template: 'actions',
-                payload: { id: 'NTM000021079' },
-                align: 'right',
-              },
-            ],
-          },
-          {
-            id: 'NTM000021080',
-            active: false,
-            data: [
-              { template: 'ticketId', payload: { id: 'NTM000021080' } },
-              {},
-              { template: 'status', payload: { status: 'active' } },
-              'Colocation',
-              24,
-              { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-              'SVUJW034781A',
-              {
-                template: 'actions',
-                payload: { id: 'NTM000021080' },
-                align: 'right',
-              },
-            ],
-          },
-          {
-            id: 'NTM000021081',
-            active: false,
-            data: [
-              { template: 'ticketId', payload: { id: 'NTM000021081' } },
-              {},
-              { template: 'status', payload: { status: 'inactive' } },
-              'Colocation',
-              25,
-              { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-              'SVUJW034781A',
-              {
-                template: 'actions',
-                payload: { id: 'NTM000021081' },
-                align: 'right',
-              },
-            ],
-          },
-          {
-            id: 'NTM000021082',
-            active: false,
-            data: [
-              { template: 'ticketId', payload: { id: 'NTM000021082' } },
-              {},
-              { template: 'status', payload: { status: 'inactive' } },
-              'Colocation',
-              26,
-              { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-              'SVUJW034781A',
-              {
-                template: 'actions',
-                payload: { id: 'NTM000021082' },
-                align: 'right',
-              },
-            ],
-          },
-          {
-            id: 'NTM000021083',
-            active: false,
-            data: [
-              { template: 'ticketId', payload: { id: 'NTM000021083' } },
-              {},
-              { template: 'status', payload: { status: 'active' } },
-              'Colocation',
-              27,
-              { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-              'SVUJW034781A',
-              {
-                template: 'actions',
-                payload: { id: 'NTM000021083' },
-                align: 'right',
-              },
-            ],
-          },
-          {
-            id: 'NTM000021084',
-            active: false,
-            data: [
-              { template: 'ticketId', payload: { id: 'NTM000021084' } },
-              {},
-              { template: 'status', payload: { status: 'active' } },
-              'Colocation',
-              28,
-              { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-              'SVUJW034781A',
-              {
-                template: 'actions',
-                payload: { id: 'NTM000021084' },
-                align: 'right',
-              },
-            ],
-          },
-          {
-            id: 'NTM000021085',
-            active: false,
-            data: [
-              { template: 'ticketId', payload: { id: 'NTM000021085' } },
-              {},
-              { template: 'status', payload: { status: 'inactive' } },
-              'Colocation',
-              29,
-              { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-              'SVUJW034781A',
-              {
-                template: 'actions',
-                payload: { id: 'NTM000021085' },
-                align: 'right',
-              },
-            ],
-          },
-          {
-            id: 'NTM000021086',
-            active: false,
-            data: [
-              { template: 'ticketId', payload: { id: 'NTM000021086' } },
-              {},
-              { template: 'status', payload: { status: 'active' } },
-              'Colocation',
-              30,
-              { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-              'SVUJW034781A',
-              {
-                template: 'actions',
-                payload: { id: 'NTM000021086' },
-                align: 'right',
-              },
-            ],
-          },
-          {
-            id: 'NTM000021087',
-            active: false,
-            data: [
-              { template: 'ticketId', payload: { id: 'NTM000021087' } },
-              {},
-              { template: 'status', payload: { status: 'active' } },
-              'Colocation',
-              31,
-              { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-              'SVUJW034781A',
-              {
-                template: 'actions',
-                payload: { id: 'NTM000021087' },
-                align: 'right',
-              },
-            ],
-          },
-          {
-            id: 'NTM000021088',
-            active: false,
-            data: [
-              { template: 'ticketId', payload: { id: 'NTM000021088' } },
-              {},
-              { template: 'status', payload: { status: 'active' } },
-              'Colocation',
-              32,
-              { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-              'SVUJW034781A',
-              {
-                template: 'actions',
-                payload: { id: 'NTM000021088' },
-                align: 'right',
-              },
-            ],
-          },
-          {
-            id: 'NTM000021089',
-            active: false,
-            data: [
-              { template: 'ticketId', payload: { id: 'NTM000021089' } },
-              {},
-              { template: 'status', payload: { status: 'inactive' } },
-              'Colocation',
-              33,
-              { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-              'SVUJW034781A',
-              {
-                template: 'actions',
-                payload: { id: 'NTM000021089' },
-                align: 'right',
-              },
-            ],
-          },
-          {
-            id: 'NTM000021090',
-            active: false,
-            data: [
-              { template: 'ticketId', payload: { id: 'NTM000021090' } },
-              {},
-              { template: 'status', payload: { status: 'inactive' } },
-              'Colocation',
-              34,
-              { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-              'SVUJW034781A',
-              {
-                template: 'actions',
-                payload: { id: 'NTM000021090' },
-                align: 'right',
-              },
-            ],
-          },
-          {
-            id: 'NTM000021091',
-            active: false,
-            data: [
-              { template: 'ticketId', payload: { id: 'NTM000021091' } },
-              {},
-              { template: 'status', payload: { status: 'active' } },
-              'Colocation',
-              35,
-              { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-              'SVUJW034781A',
-              {
-                template: 'actions',
-                payload: { id: 'NTM000021091' },
-                align: 'right',
-              },
-            ],
-          },
-          {
-            id: 'NTM000021092',
-            active: false,
-            data: [
-              { template: 'ticketId', payload: { id: 'NTM000021092' } },
-              {},
-              { template: 'status', payload: { status: 'active' } },
-              'Colocation',
-              36,
-              { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-              'SVUJW034781A',
-              {
-                template: 'actions',
-                payload: { id: 'NTM000021092' },
-                align: 'right',
-              },
-            ],
-          },
-          {
-            id: 'NTM000021093',
-            active: false,
-            data: [
-              { template: 'ticketId', payload: { id: 'NTM000021093' } },
-              {},
-              { template: 'status', payload: { status: 'inactive' } },
-              'Colocation',
-              37,
-              { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-              'SVUJW034781A',
-              {
-                template: 'actions',
-                payload: { id: 'NTM000021093' },
-                align: 'right',
-              },
-            ],
-          },
-          {
-            id: 'NTM000021094',
-            active: false,
-            data: [
-              { template: 'ticketId', payload: { id: 'NTM000021094' } },
-              {},
-              { template: 'status', payload: { status: 'inactive' } },
-              'Colocation',
-              38,
-              { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-              'SVUJW034781A',
-              {
-                template: 'actions',
-                payload: { id: 'NTM000021094' },
-                align: 'right',
-              },
-            ],
-          },
-          {
-            id: 'NTM000021095',
-            active: false,
-            data: [
-              { template: 'ticketId', payload: { id: 'NTM000021095' } },
-              {},
-              { template: 'status', payload: { status: 'active' } },
-              'Colocation',
-              39,
-              { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-              'SVUJW034781A',
-              {
-                template: 'actions',
-                payload: { id: 'NTM000021095' },
-                align: 'right',
-              },
-            ],
-          },
-          {
-            id: 'NTM000021096',
-            active: false,
-            data: [
-              { template: 'ticketId', payload: { id: 'NTM000021096' } },
-              {},
-              { template: 'status', payload: { status: 'active' } },
-              'Colocation',
-              40,
-              { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-              'SVUJW034781A',
-              {
-                template: 'actions',
-                payload: { id: 'NTM000021096' },
-                align: 'right',
-              },
-            ],
-          },
-          {
-            id: 'NTM000021097',
-            active: false,
-            data: [
-              { template: 'ticketId', payload: { id: 'NTM000021097' } },
-              {},
-              { template: 'status', payload: { status: 'active' } },
-              'Colocation',
-              41,
-              { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-              'SVUJW034781A',
-              {
-                template: 'actions',
-                payload: { id: 'NTM000021097' },
-                align: 'right',
-              },
-            ],
-          },
-          {
-            id: 'NTM000021098',
-            active: false,
-            data: [
-              { template: 'ticketId', payload: { id: 'NTM000021098' } },
-              {},
-              { template: 'status', payload: { status: 'inactive' } },
-              'Colocation',
-              42,
-              { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-              'SVUJW034781A',
-              {
-                template: 'actions',
-                payload: { id: 'NTM000021098' },
-                align: 'right',
-              },
-            ],
-          },
-          {
-            id: 'NTM000021099',
-            active: false,
-            data: [
-              { template: 'ticketId', payload: { id: 'NTM000021099' } },
-              {},
-              { template: 'status', payload: { status: 'inactive' } },
-              'Colocation',
-              43,
-              { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-              'SVUJW034781A',
-              {
-                template: 'actions',
-                payload: { id: 'NTM000021099' },
-                align: 'right',
-              },
-            ],
-          },
-          {
-            id: 'NTM000021100',
-            active: false,
-            data: [
-              { template: 'ticketId', payload: { id: 'NTM000021100' } },
-              {},
-              { template: 'status', payload: { status: 'active' } },
-              'Colocation',
-              44,
-              { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-              'SVUJW034781A',
-              {
-                template: 'actions',
-                payload: { id: 'NTM000021100' },
-                align: 'right',
-              },
-            ],
-          },
-          {
-            id: 'NTM000021101',
-            active: false,
-            data: [
-              { template: 'ticketId', payload: { id: 'NTM000021101' } },
-              {},
-              { template: 'status', payload: { status: 'inactive' } },
-              'Colocation',
-              -0.5,
-              { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-              'SVUJW034781A',
-              {
-                template: 'actions',
-                payload: { id: 'NTM000021101' },
-                align: 'right',
-              },
-            ],
-          },
-          {
-            id: 'NTM000021102',
-            active: false,
-            data: [
-              { template: 'ticketId', payload: { id: 'NTM000021102' } },
-              {},
-              { template: 'status', payload: { status: 'active' } },
-              'Colocation',
-              -1,
-              { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-              'SVUJW034781A',
-              {
-                template: 'actions',
-                payload: { id: 'NTM000021102' },
-                align: 'right',
-              },
-            ],
-          },
-        ],
-      },
-      months: ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec'],
-    };
-  },
-})
-export default class DataTableView extends Vue {
-  mounted() {
-    setTimeout(() => {
-      const newData = [
-        { template: 'ticketId', payload: { id: 'NTM000021063' } },
-        {
-          template: 'icon',
-          payload: { icon: 'circle-check', color: 'success' },
-        },
-        { template: 'status', payload: { status: 'active' } },
-        'Colocation',
-        0,
-        { template: 'date', payload: { date: new Date('04/05/2018 8:00 AM') } },
-        'SVUJW034781A',
-        {
-          template: 'actions',
-          payload: { id: 'NTM000021063' },
-          align: 'right',
-        },
-      ];
-      this.$data.table = {
-        ...this.$data.table,
-        body: this.$data.table.body.map((row: DataTableRow, index: number) =>
-          index === 0
-            ? {
-                ...row,
-                nestedContent: {
-                  value: `Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam,
-              quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur.
-              quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur.
-              quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur.
-              Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum."`,
-                },
-                data: newData,
-              }
-            : row
-        ),
-      };
-    }, 5000);
-  }
-}
-</script>
-
-<style lang="scss"></style>
